refactor(paste): extract menu fallback orchestration from paste service
- split menu fallback decision/execution flow into pastemenufallbackcoordinator while keeping paste service focused on top-level orchestration, clipboard restore timing, and failure recovery lifecycle. preserve existing paste service static helper contract via delegation wrappers (decision helpers + electron detection) to avoid test/api churn.

- wire coordinator into xcode project sources and update codemap/engineering docs to reflect responsibilities. no policy changes intended; structural refactor only.
